# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginIn": {
        "properties": {
          "password": {
            "maxLength": 256,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "MemberChoice": {
        "properties": {
          "member_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          }
        },
        "required": [
          "member_id"
        ],
        "title": "MemberChoice",
        "type": "object"
      },
...
      "MemberOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "marker_color": {
            "enum": [
              "basil",
              "tomato",
              "carrot",
              "eggplant",
              "beet",
              "olive",
              "cocoa",
              "plum"
            ],
            "title": "Marker Color",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "marker_color"
        ],
        "title": "MemberOut",
        "type": "object"
      },
...
      "SessionOut": {
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "household_name": {
            "title": "Household Name",
            "type": "string"
          },
          "member": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemberOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberOut"
            },
            "title": "Members",
            "type": "array"
          }
        },
        "required": [
          "device_id",
          "member",
          "members",
          "household_name"
        ],
        "title": "SessionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Dinner Bell",
    "version": "0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/backups/{name}": {
      "get": {
        "operationId": "download_backup_api_admin_backups__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Backup",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/api/auth/devices/{device_id}": {
      "delete": {
        "operationId": "sign_out_device_api_auth_devices__device_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign Out Device",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/member": {
      "put": {
        "operationId": "choose_member_api_auth_member_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberChoice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Choose Member",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}